move base.Options to pebble.Options
Now that `base.Options` is only used by the `pebble` package we can
replace the `pebble.Options` type alias with the actual type.
