+++ internal/base/options.go
@@ -4,15 +4,11 @@
 
 package base
 
-import (
-	"bytes"
-	"errors"
-	"fmt"
-	"strconv"
-	"strings"
-
-	"github.com/cockroachdb/pebble/internal/cache"
-	"github.com/cockroachdb/pebble/vfs"
+// SSTable block defaults.
+const (
+	DefaultBlockRestartInterval = 16
+	DefaultBlockSize            = 4096
+	DefaultBlockSizeThreshold   = 90
 )
 
 // Compression is the per-block compression algorithm to use.
@@ -92,13 +88,6 @@ type FilterPolicy interface {
 	NewWriter(ftype FilterType) FilterWriter
 }
 
-func filterPolicyName(p FilterPolicy) string {
-	if p == nil {
-		return "none"
-	}
-	return p.Name()
-}
-
 // TableFormat specifies the format version for sstables. The legacy LevelDB
 // format is format version 0.
 type TableFormat uint32
@@ -130,632 +119,4 @@ type TablePropertyCollector interface {
 
 	// The name of the property collector.
 	Name() string
-}
-
-// LevelOptions holds the optional per-level parameters.
-type LevelOptions struct {
-	// BlockRestartInterval is the number of keys between restart points
-	// for delta encoding of keys.
-	//
-	// The default value is 16.
-	BlockRestartInterval int
-
-	// BlockSize is the target uncompressed size in bytes of each table block.
-	//
-	// The default value is 4096.
-	BlockSize int
-
-	// BlockSizeThreshold finishes a block if the block size is larger than the
-	// specified percentage of the target block size and adding the next entry
-	// would cause the block to be larger than the target block size.
-	//
-	// The default value is 90
-	BlockSizeThreshold int
-
-	// Compression defines the per-block compression to use.
-	//
-	// The default value (DefaultCompression) uses snappy compression.
-	Compression Compression
-
-	// FilterPolicy defines a filter algorithm (such as a Bloom filter) that can
-	// reduce disk reads for Get calls.
-	//
-	// One such implementation is bloom.FilterPolicy(10) from the pebble/bloom
-	// package.
-	//
-	// The default value means to use no filter.
-	FilterPolicy FilterPolicy
-
-	// FilterType defines whether an existing filter policy is applied at a
-	// block-level or table-level. Block-level filters use less memory to create,
-	// but are slower to access as a check for the key in the index must first be
-	// performed to locate the filter block. A table-level filter will require
-	// memory proportional to the number of keys in an sstable to create, but
-	// avoids the index lookup when determining if a key is present. Table-level
-	// filters should be preferred except under constrained memory situations.
-	FilterType FilterType
-
-	// IndexBlockSize is the target uncompressed size in bytes of each index
-	// block. When the index block size is larger than this target, two-level
-	// indexes are automatically enabled. Setting this option to a large value
-	// (such as math.MaxInt32) disables the automatic creation of two-level
-	// indexes.
-	//
-	// The default value is the value of BlockSize.
-	IndexBlockSize int
-
-	// The target file size for the level.
-	TargetFileSize int64
-}
-
-// EnsureDefaults ensures that the default values for all of the options have
-// been initialized. It is valid to call EnsureDefaults on a nil receiver. A
-// non-nil result will always be returned.
-func (o *LevelOptions) EnsureDefaults() *LevelOptions {
-	if o == nil {
-		o = &LevelOptions{}
-	}
-	if o.BlockRestartInterval <= 0 {
-		o.BlockRestartInterval = 16
-	}
-	if o.BlockSize <= 0 {
-		o.BlockSize = 4096
-	}
-	if o.BlockSizeThreshold <= 0 {
-		o.BlockSizeThreshold = 90
-	}
-	if o.Compression <= DefaultCompression || o.Compression >= NCompression {
-		o.Compression = SnappyCompression
-	}
-	if o.IndexBlockSize <= 0 {
-		o.IndexBlockSize = o.BlockSize
-	}
-	if o.TargetFileSize <= 0 {
-		o.TargetFileSize = 2 << 20 // 2 MB
-	}
-	return o
-}
-
-// Options holds the optional parameters for configuring pebble. These options
-// apply to the DB at large; per-query options are defined by the IterOptions
-// and WriteOptions types.
-type Options struct {
-	// Sync sstables and the WAL periodically in order to smooth out writes to
-	// disk. This option does not provide any persistency guarantee, but is used
-	// to avoid latency spikes if the OS automatically decides to write out a
-	// large chunk of dirty filesystem buffers.
-	//
-	// The default value is 512KB.
-	BytesPerSync int
-
-	// Cache is used to cache uncompressed blocks from sstables.
-	//
-	// The default cache size is 8 MB.
-	Cache *cache.Cache
-
-	// Cleaner cleans obsolete files.
-	//
-	// The default cleaner uses the DeleteCleaner.
-	Cleaner Cleaner
-
-	// Comparer defines a total ordering over the space of []byte keys: a 'less
-	// than' relationship. The same comparison algorithm must be used for reads
-	// and writes over the lifetime of the DB.
-	//
-	// The default value uses the same ordering as bytes.Compare.
-	Comparer *Comparer
-
-	// Disable the write-ahead log (WAL). Disabling the write-ahead log prohibits
-	// crash recovery, but can improve performance if crash recovery is not
-	// needed (e.g. when only temporary state is being stored in the database).
-	//
-	// TODO(peter): untested
-	DisableWAL bool
-
-	// ErrorIfDBExists is whether it is an error if the database already exists.
-	//
-	// The default value is false.
-	ErrorIfDBExists bool
-
-	// EventListener provides hooks to listening to significant DB events such as
-	// flushes, compactions, and table deletion.
-	EventListener EventListener
-
-	// Filters is a map from filter policy name to filter policy. It is used for
-	// debugging tools which may be used on multiple databases configured with
-	// different filter policies. It is not necessary to populate this filters
-	// map during normal usage of a DB.
-	Filters map[string]FilterPolicy
-
-	// FS provides the interface for persistent file storage.
-	//
-	// The default value uses the underlying operating system's file system.
-	FS vfs.FS
-
-	// The number of files necessary to trigger an L0 compaction.
-	L0CompactionThreshold int
-
-	// Hard limit on the number of L0 files. Writes are stopped when this
-	// threshold is reached.
-	L0StopWritesThreshold int
-
-	// The maximum number of bytes for LBase. The base level is the level which
-	// L0 is compacted into. The base level is determined dynamically based on
-	// the existing data in the LSM. The maximum number of bytes for other levels
-	// is computed dynamically based on the base level's maximum size. When the
-	// maximum number of bytes for a level is exceeded, compaction is requested.
-	LBaseMaxBytes int64
-
-	// Per-level options. Options for at least one level must be specified. The
-	// options for the last level are used for all subsequent levels.
-	Levels []LevelOptions
-
-	// Logger used to write log messages.
-	//
-	// The default logger uses the Go standard library log package.
-	Logger Logger
-
-	// MaxManifestFileSize is the maximum size the MANIFEST file is allowed to
-	// become. When the MANIFEST exceeds this size it is rolled over and a new
-	// MANIFEST is created.
-	MaxManifestFileSize int64
-
-	// MaxOpenFiles is a soft limit on the number of open files that can be
-	// used by the DB.
-	//
-	// The default value is 1000.
-	MaxOpenFiles int
-
-	// The size of a MemTable. Note that more than one MemTable can be in
-	// existence since flushing a MemTable involves creating a new one and
-	// writing the contents of the old one in the
-	// background. MemTableStopWritesThreshold places a hard limit on the number
-	// of MemTables allowed at once.
-	MemTableSize int
-
-	// Hard limit on the number of MemTables. Writes are stopped when this number
-	// is reached. This value should be at least 2 or writes will stop whenever
-	// the MemTable is being flushed.
-	MemTableStopWritesThreshold int
-
-	// Merger defines the associative merge operation to use for merging values
-	// written with {Batch,DB}.Merge.
-	//
-	// The default merger concatenates values.
-	Merger *Merger
-
-	// MinCompactionRate sets the minimum rate at which compactions occur. The
-	// default is 4 MB/s.
-	MinCompactionRate int
-
-	// MinFlushRate sets the minimum rate at which the MemTables are flushed. The
-	// default is 1 MB/s.
-	MinFlushRate int
-
-	// ReadOnly indicates that the DB should be opened in read-only mode. Writes
-	// to the DB will return an error, background compactions are disabled, and
-	// the flush that normally occurs after replaying the WAL at startup is
-	// disabled.
-	ReadOnly bool
-
-	// TableFormat specifies the format version for writing sstables. The default
-	// is TableFormatRocksDBv2 which creates RocksDB compatible sstables. Use
-	// TableFormatLevelDB to create LevelDB compatible sstable which can be used
-	// by a wider range of tools and libraries.
-	TableFormat TableFormat
-
-	// TablePropertyCollectors is a list of TablePropertyCollector creation
-	// functions. A new TablePropertyCollector is created for each sstable built
-	// and lives for the lifetime of the table.
-	TablePropertyCollectors []func() TablePropertyCollector
-
-	// WALDir specifies the directory to store write-ahead logs (WALs) in. If
-	// empty (the default), WALs will be stored in the same directory as sstables
-	// (i.e. the directory passed to pebble.Open).
-	WALDir string
-}
-
-// EnsureDefaults ensures that the default values for all options are set if a
-// valid value was not already specified. Returns the new options.
-func (o *Options) EnsureDefaults() *Options {
-	if o == nil {
-		o = &Options{}
-	}
-	if o.BytesPerSync <= 0 {
-		o.BytesPerSync = 512 << 10 // 512 KB
-	}
-	if o.Cache == nil {
-		o.Cache = cache.New(8 << 20) // 8 MB
-	}
-	if o.Cleaner == nil {
-		o.Cleaner = DeleteCleaner{}
-	}
-	if o.Comparer == nil {
-		o.Comparer = DefaultComparer
-	}
-	if o.FS == nil {
-		o.FS = vfs.Default
-	}
-	if o.L0CompactionThreshold <= 0 {
-		o.L0CompactionThreshold = 4
-	}
-	if o.L0StopWritesThreshold <= 0 {
-		o.L0StopWritesThreshold = 12
-	}
-	if o.LBaseMaxBytes <= 0 {
-		o.LBaseMaxBytes = 64 << 20 // 64 MB
-	}
-	if o.Levels == nil {
-		o.Levels = make([]LevelOptions, 1)
-		for i := range o.Levels {
-			if i > 0 {
-				l := &o.Levels[i]
-				if l.TargetFileSize <= 0 {
-					l.TargetFileSize = o.Levels[i-1].TargetFileSize * 2
-				}
-			}
-			o.Levels[i].EnsureDefaults()
-		}
-	} else {
-		for i := range o.Levels {
-			o.Levels[i].EnsureDefaults()
-		}
-	}
-	if o.Logger == nil {
-		o.Logger = defaultLogger{}
-	}
-	o.EventListener.EnsureDefaults(o.Logger)
-	if o.MaxManifestFileSize == 0 {
-		o.MaxManifestFileSize = 128 << 20 // 128 MB
-	}
-	if o.MaxOpenFiles == 0 {
-		o.MaxOpenFiles = 1000
-	}
-	if o.MemTableSize <= 0 {
-		o.MemTableSize = 4 << 20
-	}
-	if o.MemTableStopWritesThreshold <= 0 {
-		o.MemTableStopWritesThreshold = 2
-	}
-	if o.Merger == nil {
-		o.Merger = DefaultMerger
-	}
-	if o.MinCompactionRate == 0 {
-		o.MinCompactionRate = 4 << 20 // 4 MB/s
-	}
-	if o.MinFlushRate == 0 {
-		o.MinFlushRate = 1 << 20 // 1 MB/s
-	}
-
-	o.initMaps()
-	return o
-}
-
-// initMaps initializes the Comparers, Filters, and Mergers maps.
-func (o *Options) initMaps() {
-	for i := range o.Levels {
-		l := &o.Levels[i]
-		if l.FilterPolicy != nil {
-			if o.Filters == nil {
-				o.Filters = make(map[string]FilterPolicy)
-			}
-			name := l.FilterPolicy.Name()
-			if _, ok := o.Filters[name]; !ok {
-				o.Filters[name] = l.FilterPolicy
-			}
-		}
-	}
-}
-
-// Level returns the LevelOptions for the specified level.
-func (o *Options) Level(level int) LevelOptions {
-	if level < len(o.Levels) {
-		return o.Levels[level]
-	}
-	n := len(o.Levels) - 1
-	l := o.Levels[n]
-	for i := n; i < level; i++ {
-		l.TargetFileSize *= 2
-	}
-	return l
-}
-
-// Clone creates a shallow-copy of the supplied options.
-func (o *Options) Clone() *Options {
-	n := &Options{}
-	if o != nil {
-		*n = *o
-	}
-	return n
-}
-
-func (o *Options) String() string {
-	var buf bytes.Buffer
-
-	fmt.Fprintf(&buf, "[Version]\n")
-	fmt.Fprintf(&buf, "  pebble_version=0.1\n")
-	fmt.Fprintf(&buf, "\n")
-	fmt.Fprintf(&buf, "[Options]\n")
-	fmt.Fprintf(&buf, "  bytes_per_sync=%d\n", o.BytesPerSync)
-	fmt.Fprintf(&buf, "  cache_size=%d\n", o.Cache.MaxSize())
-	fmt.Fprintf(&buf, "  cleaner=%s\n", o.Cleaner)
-	fmt.Fprintf(&buf, "  comparer=%s\n", o.Comparer.Name)
-	fmt.Fprintf(&buf, "  disable_wal=%t\n", o.DisableWAL)
-	fmt.Fprintf(&buf, "  l0_compaction_threshold=%d\n", o.L0CompactionThreshold)
-	fmt.Fprintf(&buf, "  l0_stop_writes_threshold=%d\n", o.L0StopWritesThreshold)
-	fmt.Fprintf(&buf, "  lbase_max_bytes=%d\n", o.LBaseMaxBytes)
-	fmt.Fprintf(&buf, "  max_manifest_file_size=%d\n", o.MaxManifestFileSize)
-	fmt.Fprintf(&buf, "  max_open_files=%d\n", o.MaxOpenFiles)
-	fmt.Fprintf(&buf, "  mem_table_size=%d\n", o.MemTableSize)
-	fmt.Fprintf(&buf, "  mem_table_stop_writes_threshold=%d\n", o.MemTableStopWritesThreshold)
-	fmt.Fprintf(&buf, "  min_compaction_rate=%d\n", o.MinCompactionRate)
-	fmt.Fprintf(&buf, "  min_flush_rate=%d\n", o.MinFlushRate)
-	fmt.Fprintf(&buf, "  merger=%s\n", o.Merger.Name)
-	fmt.Fprintf(&buf, "  table_property_collectors=[")
-	for i := range o.TablePropertyCollectors {
-		if i > 0 {
-			fmt.Fprintf(&buf, ",")
-		}
-		// NB: This creates a new TablePropertyCollector, but Options.String() is
-		// called rarely so the overhead of doing so is not consequential.
-		fmt.Fprintf(&buf, "%s", o.TablePropertyCollectors[i]().Name())
-	}
-	fmt.Fprintf(&buf, "]\n")
-	fmt.Fprintf(&buf, "  wal_dir=%s\n", o.WALDir)
-
-	for i := range o.Levels {
-		l := &o.Levels[i]
-		fmt.Fprintf(&buf, "\n")
-		fmt.Fprintf(&buf, "[Level \"%d\"]\n", i)
-		fmt.Fprintf(&buf, "  block_restart_interval=%d\n", l.BlockRestartInterval)
-		fmt.Fprintf(&buf, "  block_size=%d\n", l.BlockSize)
-		fmt.Fprintf(&buf, "  compression=%s\n", l.Compression)
-		fmt.Fprintf(&buf, "  filter_policy=%s\n", filterPolicyName(l.FilterPolicy))
-		fmt.Fprintf(&buf, "  filter_type=%s\n", l.FilterType)
-		fmt.Fprintf(&buf, "  index_block_size=%d\n", l.IndexBlockSize)
-		fmt.Fprintf(&buf, "  target_file_size=%d\n", l.TargetFileSize)
-	}
-
-	return buf.String()
-}
-
-func parseOptions(s string, fn func(section, key, value string) error) error {
-	var section string
-	for _, line := range strings.Split(s, "\n") {
-		line = strings.TrimSpace(line)
-		if len(line) == 0 {
-			// Skip blank lines.
-			continue
-		}
-		if line[0] == ';' || line[0] == '#' {
-			// Skip comments.
-			continue
-		}
-		n := len(line)
-		if line[0] == '[' && line[n-1] == ']' {
-			// Parse section.
-			section = line[1 : n-1]
-			continue
-		}
-
-		pos := strings.Index(line, "=")
-		if pos < 0 {
-			return fmt.Errorf("pebble: invalid key=value syntax: %s", line)
-		}
-
-		key := strings.TrimSpace(line[:pos])
-		value := strings.TrimSpace(line[pos+1:])
-		if err := fn(section, key, value); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// ParseHooks contains callbacks to create options fields which can have
-// user-defined implementations.
-type ParseHooks struct {
-	NewCleaner      func(name string) (Cleaner, error)
-	NewComparer     func(name string) (*Comparer, error)
-	NewFilterPolicy func(name string) (FilterPolicy, error)
-	NewMerger       func(name string) (*Merger, error)
-}
-
-// Parse parses the options from the specified string. Note that certain
-// options cannot be parsed into populated fields. For example, comparer and
-// merger.
-func (o *Options) Parse(s string, hooks *ParseHooks) error {
-	return parseOptions(s, func(section, key, value string) error {
-		switch {
-		case section == "Version":
-			switch key {
-			case "pebble_version":
-			default:
-				return fmt.Errorf("pebble: unknown option: %s.%s", section, key)
-			}
-			return nil
-
-		case section == "Options":
-			var err error
-			switch key {
-			case "bytes_per_sync":
-				o.BytesPerSync, err = strconv.Atoi(value)
-			case "cache_size":
-				n, err := strconv.ParseInt(value, 10, 64)
-				if err == nil {
-					o.Cache = cache.New(n)
-				}
-			case "cleaner":
-				switch value {
-				case "archive":
-					o.Cleaner = ArchiveCleaner{}
-				case "delete":
-					o.Cleaner = DeleteCleaner{}
-				default:
-					if hooks != nil && hooks.NewCleaner != nil {
-						o.Cleaner, err = hooks.NewCleaner(value)
-					}
-				}
-			case "comparer":
-				switch value {
-				case "leveldb.BytewiseComparator":
-					o.Comparer = DefaultComparer
-				default:
-					if hooks != nil && hooks.NewComparer != nil {
-						o.Comparer, err = hooks.NewComparer(value)
-					}
-				}
-			case "disable_wal":
-				o.DisableWAL, err = strconv.ParseBool(value)
-			case "l0_compaction_threshold":
-				o.L0CompactionThreshold, err = strconv.Atoi(value)
-			case "l0_stop_writes_threshold":
-				o.L0StopWritesThreshold, err = strconv.Atoi(value)
-			case "lbase_max_bytes":
-				o.LBaseMaxBytes, err = strconv.ParseInt(value, 10, 64)
-			case "max_manifest_file_size":
-				o.MaxManifestFileSize, err = strconv.ParseInt(value, 10, 64)
-			case "max_open_files":
-				o.MaxOpenFiles, err = strconv.Atoi(value)
-			case "mem_table_size":
-				o.MemTableSize, err = strconv.Atoi(value)
-			case "mem_table_stop_writes_threshold":
-				o.MemTableStopWritesThreshold, err = strconv.Atoi(value)
-			case "min_compaction_rate":
-				o.MinCompactionRate, err = strconv.Atoi(value)
-			case "min_flush_rate":
-				o.MinFlushRate, err = strconv.Atoi(value)
-			case "merger":
-				switch value {
-				case "pebble.concatenate":
-					o.Merger = DefaultMerger
-				default:
-					if hooks != nil && hooks.NewMerger != nil {
-						o.Merger, err = hooks.NewMerger(value)
-					}
-				}
-			case "table_format":
-				switch value {
-				case "leveldb":
-					o.TableFormat = TableFormatLevelDB
-				case "rocksdbv2":
-					o.TableFormat = TableFormatRocksDBv2
-				default:
-					return fmt.Errorf("pebble: unknown table format: %q", value)
-				}
-			case "table_property_collectors":
-				// TODO(peter): set o.TablePropertyCollectors
-			case "wal_dir":
-				o.WALDir = value
-			default:
-				return fmt.Errorf("pebble: unknown option: %s.%s", section, key)
-			}
-			return err
-
-		case strings.HasPrefix(section, "Level "):
-			var index int
-			if n, err := fmt.Sscanf(section, `Level "%d"`, &index); err != nil {
-				return err
-			} else if n != 1 {
-				return fmt.Errorf("pebble: unknown section: %q", section)
-			}
-
-			if len(o.Levels) <= index {
-				newLevels := make([]LevelOptions, index+1)
-				copy(newLevels, o.Levels)
-				o.Levels = newLevels
-			}
-			l := &o.Levels[index]
-
-			var err error
-			switch key {
-			case "block_restart_interval":
-				l.BlockRestartInterval, err = strconv.Atoi(value)
-			case "block_size":
-				l.BlockSize, err = strconv.Atoi(value)
-			case "compression":
-				switch value {
-				case "Default":
-					l.Compression = DefaultCompression
-				case "NoCompression":
-					l.Compression = NoCompression
-				case "Snappy":
-					l.Compression = SnappyCompression
-				default:
-					return fmt.Errorf("pebble: unknown compression: %q", value)
-				}
-			case "filter_policy":
-				if hooks != nil && hooks.NewFilterPolicy != nil {
-					l.FilterPolicy, err = hooks.NewFilterPolicy(value)
-				}
-			case "filter_type":
-				switch value {
-				case "table":
-					l.FilterType = TableFilter
-				default:
-					return fmt.Errorf("pebble: unknown filter type: %q", value)
-				}
-			case "index_block_size":
-				l.IndexBlockSize, err = strconv.Atoi(value)
-			case "target_file_size":
-				l.TargetFileSize, err = strconv.ParseInt(value, 10, 64)
-			default:
-				return fmt.Errorf("pebble: unknown option: %s.%s", section, key)
-			}
-			return err
-		}
-		return fmt.Errorf("pebble: unknown section: %q", section)
-	})
-}
-
-// Check verifies the options are compatible with the previous options
-// serialized by Options.String(). For example, the Comparer and Merger must be
-// the same, or data will not be able to be properly read from the DB.
-func (o *Options) Check(s string) error {
-	return parseOptions(s, func(section, key, value string) error {
-		// RocksDB uses a similar (INI-style) syntax for the OPTIONS file, but
-		// different section names and keys. The "CFOptions ..." paths below are
-		// the RocksDB versions.
-		switch section + "." + key {
-		case "Options.comparer", `CFOptions "default".comparator`:
-			if value != o.Comparer.Name {
-				return fmt.Errorf("pebble: comparer name from file %q != comparer name from options %q",
-					value, o.Comparer.Name)
-			}
-		case "Options.merger", `CFOptions "default".merge_operator`:
-			// RocksDB allows the merge operator to be unspecified, in which case it
-			// shows up as "nullptr".
-			if value != "nullptr" && value != o.Merger.Name {
-				return fmt.Errorf("pebble: merger name from file %q != merger name from options %q",
-					value, o.Merger.Name)
-			}
-		}
-		return nil
-	})
-}
-
-// Validate verifies that the options are mutually consistent. For example,
-// L0StopWritesThreshold must be >= L0CompactionThreshold, otherwise a write
-// stall would persist indefinitely.
-func (o *Options) Validate() error {
-	// Note that we can presume Options.EnsureDefaults has been called, so there
-	// is no need to check for zero values.
-
-	var buf strings.Builder
-	if o.L0StopWritesThreshold < o.L0CompactionThreshold {
-		fmt.Fprintf(&buf, "L0StopWritesThreshold (%d) must be >= L0CompactionThreshold (%d)\n",
-			o.L0StopWritesThreshold, o.L0CompactionThreshold)
-	}
-	if o.MemTableStopWritesThreshold < 2 {
-		fmt.Fprintf(&buf, "MemTableStopWritesThreshold (%d) must be >= 2\n",
-			o.MemTableStopWritesThreshold)
-	}
-	switch o.TableFormat {
-	case TableFormatLevelDB:
-		fmt.Fprintf(&buf, "TableFormatLevelDB not supported for DB\n")
-	}
-	if buf.Len() == 0 {
-		return nil
-	}
-	return errors.New(buf.String())
 }
